Parsers: Init of .obj parser

+++ src/Support/MacOS/Window.h
@@ -16,7 +16,7 @@ public:
     void set_key_up_callback(void (*key_callback)(int, int));
     void set_key_down_callback(void (*key_callback)(int, int));
     void set_mouse_down_callback(void (*mouse_down_callback)(int, int));
-    void set_mouse_move_callback(void (*mouse_move_callback)(float, float));
+    void set_mouse_move_callback(void (*mouse_move_callback)(int, int));
 
     static void Run();
